CHAD-17062: zigbee-bed lazy loading subdrivers
