+++ drivers/SmartThings/zigbee-bed/src/shus-mattress/fingerprints.lua
@@ -1,0 +1,8 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local FINGERPRINTS = {
+  { mfr = "SHUS", model = "SX-1" }
+}
+
+return FINGERPRINTS
